backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
python-docx>=1.0.0
pydantic>=2.0.0
